Rename all tag files to allergy
