--- src/main/java/seedu/address/model/tag/Tag.java
@@ -1,75 +1,0 @@
-package seedu.address.model.tag;
-
-import static java.util.Objects.requireNonNull;
-import static seedu.address.commons.util.AppUtil.checkArgument;
-
-/**
- * Represents a Tag in the address book.
- * Guarantees: immutable; name is valid as declared in {@link #isValidTagName(String)}
- */
-public class Tag implements Comparable<Tag> {
-
-    public static final String MESSAGE_CONSTRAINTS = "Tags names should be alphanumeric and not exceed 30 characters";
-    public static final String VALIDATION_REGEX = "\\p{Alnum}+(\\s\\p{Alnum}+)*";
-
-    public final String tagName;
-
-    /**
-     * Constructs a {@code Tag}.
-     *
-     * @param tagName A valid tag name.
-     */
-    public Tag(String tagName) {
-        requireNonNull(tagName);
-        checkArgument(isValidTagName(tagName), MESSAGE_CONSTRAINTS);
-        this.tagName = tagName;
-    }
-
-    /**
-     * Returns true if a given string is a valid tag name.
-     */
-    public static boolean isValidTagName(String test) {
-        return test.matches(VALIDATION_REGEX) && test.length() <= 30;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof Tag)) {
-            return false;
-        }
-
-        Tag otherTag = (Tag) other;
-        return tagName.equals(otherTag.tagName);
-    }
-
-    @Override
-    public int hashCode() {
-        return tagName.hashCode();
-    }
-
-    /**
-     * Format state as text for viewing.
-     */
-    public String toString() {
-        return tagName;
-    }
-
-    /**
-     * Compares this {@code Tag} object with the specified {@code Tag} object for order.
-     * Returns a negative integer, zero, or a positive integer if this {@code Tag}
-     * is lexicographically less than, equal to, or greater than the specified {@code Tag}.
-     *
-     * @param other the {@code Tag} object to be compared.
-     * @return a negative integer, zero, or a positive integer as this {@code Tag}
-     *         is less than, equal to, or greater than the specified {@code Tag}.
-     */
-    @Override
-    public int compareTo(Tag other) {
-        return this.tagName.compareTo(other.tagName);
-    }
-}
