+++ src/test/java/seedu/address/model/allergy/AllergyTest.java
